chore: upgrade dependencies major version
