+++ package.json
@@ -4,7 +4,6 @@
 	"engines": {
 		"node": "24.x"
 	},
-	"type": "module",
 	"private": true,
 	"description": "ARK Vault",
 	"homepage": "./",
